Update ufspj.com
I added some extra padding to the article text. I also added a solid light grey border to the section, in the same color as the links to tie the look together. I added media queries for a phone format. I also added more details to the Twitter, like text-align and padding. I added a hover link color for the navigation and header links, in the same navy color as the background.